requirements.txt
taipy==2.1
numpy==1.22.4
tensorflow-cpu==2.9.1
Pillow
gunicorn
opencv-python
wget
